# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controller.createProductRequest:
    properties:
      category_id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      unit_price:
        type: number
    required:
    - category_id
    - name
    - quantity
    - sku
    - unit_price
    type: object
...
  controller.newOrderRequest:
    properties:
      address:
        type: string
      card_cvv:
        type: integer
      card_exp:
        type: string
      card_number:
        type: string
      name:
        type: string
      phone_number:
        type: string
    required:
    - address
    - card_cvv
    - card_exp
    - card_number
    - name
    - phone_number
    type: object
  controller.orderResponse:
    properties:
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/entity.OrderItem'
        type: array
    type: object
  controller.productResponse:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      unit_price:
        type: number
    type: object
...
  controller.response:
    properties:
      message:
        example: message
        type: string
    type: object
...
  entity.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/entity.Product'
        type: array
      updated_at:
        type: string
    type: object
  entity.OrderItem:
    properties:
      created_at:
        type: string
      order_id:
        type: string
      product:
        $ref: '#/definitions/entity.Product'
      product_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Product:
    properties:
      category:
        $ref: '#/definitions/entity.Category'
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      unit_price:
        type: number
      updated_at:
        type: string
    type: object
  pagination.Pages:
    properties:
      items: {}
      page:
        type: integer
      pageCount:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: An API for store management
  title: Beste Tufan Store API
  version: "1.0"
paths:
...
  /order:
    get:
      consumes:
      - application/json
      description: Returns user's orders.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.orderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - Bearer: []
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Creates an order with basket items.
      parameters:
      - description: New Order Model
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.newOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - Bearer: []
      tags:
      - Order
  /order/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancels an order.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - Bearer: []
      tags:
      - Order
  /product:
    delete:
      consumes:
      - application/json
      description: Deletes a product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Returns all products with pagination.
      parameters:
      - description: Page Index
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - Bearer: []
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Creates a new product.
      parameters:
      - description: Create Product Model
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.createProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.productResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - Bearer: []
      tags:
      - Product
  /product/{id}:
    get:
      consumes:
      - application/json
      description: Returns one product by id.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.productResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - Bearer: []
      tags:
      - Product
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"